Make protobuf optional (default on)
