+++ include/zq/zq.hpp
@@ -3,5 +3,9 @@
 #include "context.hpp"
 
 #include "message.hpp"
+#ifdef ZQ_PROTO
+  #include "message_proto.hpp"
+#endif
+
 #include "socket.hpp"
 #include "zflags.hpp"
