+++ tests/commu/hello_test.cpp
@@ -1,10 +1,8 @@
 #include <doctest/doctest.h>
 #include <zq/zq.hpp>
-// #include <zq/typename.hpp>
 #include <fmt/format.h>
 #include <chrono>
 #include <thread>
-#include "pingpong.pb.h"
 
 namespace {
   // startup times on Windows are a problem, they take too long,
@@ -80,10 +78,11 @@ SCENARIO("Make a hello world send receive call") {
   }
 }
 
+
 SCENARIO("Make a hello world send receive call") {
   auto context = zq::mk_context();
 
-  GIVEN("a request and a reply socket with proto messages") {
+  GIVEN("a request and a reply socket int message") {
     // auto client = context->connect(zq::SocketType::REQ,
     // "tcp://localhost:5555"); auto server = context->bind(zq::SocketType::REP,
     // "tcp://localhost:5555");
@@ -93,42 +92,39 @@ SCENARIO("Make a hello world send receive call") {
     REQUIRE(server);
 
     WHEN("requesting a ping reply") {
-      zq::proto::Ping ping;
-      ping.set_id(1);
-      ping.set_msg("Hi");
-      ;
-      auto tm = zq::typed_message("Hello world");
-      auto res = client->send(zq::typed_message(ping));
+
+      std::string hello_str = "Hello world";
+      auto tm = zq::typed_message(hello_str);
+      auto res = client->send(tm);
       REQUIRE(res);
 
       THEN("it's possible to receive and restore the message") {
         auto request = server->await(await_time);
         REQUIRE(request);
         REQUIRE(request.value());
-        auto restored = zq::restore_as<zq::proto::Ping>(*request.value());
+        auto restored = zq::restore_as<std::string>(*request.value());
         if (!restored) {
           MESSAGE("error: ", restored.error().what());
         }
         REQUIRE(restored);
-        REQUIRE_EQ(restored->id(), 1);
-        REQUIRE_EQ(restored->msg(), "Hi");
+        REQUIRE_EQ(*restored, hello_str);
+
         AND_THEN("it's possible to send and receive a reply") {
-          zq::proto::Pong pong;
-          pong.set_id(1);
-          pong.set_reply("Hi there");
-          auto send_rc = server->send(zq::typed_message(pong));
+          std::string reply_str = "Hi there";
+          auto send_rc = server->send(zq::typed_message(reply_str));
           REQUIRE(send_rc);
           auto reply = client->await(await_time);
           REQUIRE(reply);
           REQUIRE(reply.value());
-          auto restored_reply = zq::restore_as<zq::proto::Pong>(*reply.value());
-          REQUIRE_EQ(restored_reply->id(), 1);
-          REQUIRE_EQ(restored_reply->reply(), "Hi there");
+          auto restored_reply = zq::restore_as<std::string>(*reply.value());
+          REQUIRE(restored_reply);
+          REQUIRE_EQ(*restored_reply, reply_str);
         }
       }
     }
   }
 }
+
 
 SCENARIO(
     "Trying to create a socket without valid context returns error message") {
